Silence warnings regarding possible loss of precision converting from integer to floating-point types. Removed unnecessary conversion to float in FTimecode::FromFrameNumber() Minimize loss of precision in FThreadHeartBeatClock::Seconds() for long-running titles FGenericPlatformMath::FRand() now truncates Rand() to 24 bits or RAND_MAX, whichever is smaller.  This prevents values closer to 1.0 being higher probability due to loss of precision. #rb luke.thatcher, eric.mcdaniel, brian.white #jira none
#ROBOMERGE-SOURCE: CL 14574582 in //UE4/Release-4.26/...
#ROBOMERGE-BOT: RELEASE (Release-4.26 -> Release-Engine-Staging) (v750-14555174)

[CL 14574596 by jeff newquist in Release-Engine-Staging branch]

+++ Engine/Source/Runtime/Core/Private/Containers/StackTracker.cpp
@@ -166,7 +166,7 @@ void FStackTracker::DumpStackTraces(int32 StackThreshold, FOutputDevice& Ar, flo
 	}
 	else
 	{
-		Ar.Logf(TEXT("Captured %i unique callstacks totalling %i function calls over %i frames, averaging %5.2f calls/frame, Avg Per Frame"), SortedCallStacks.Num(), (int32)TotalStackCount, FramesCaptured, (float) TotalStackCount / FramesCaptured);
+		Ar.Logf(TEXT("Captured %i unique callstacks totalling %i function calls over %i frames, averaging %5.2f calls/frame, Avg Per Frame"), SortedCallStacks.Num(), (int32)TotalStackCount, FramesCaptured, (float)TotalStackCount / (float)FramesCaptured);
 	}
 
 	// Iterate over each callstack and write out info in human readable form in CSV format
